+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## v0.1.8 - 2026-08-13
+
 ### Added
 
 - Scroll to letter when user clicks to view, on landing page